+++ api/core/main.py
@@ -2,10 +2,11 @@ from core import database
 from core import settings as proj_settings
 from fastapi import FastAPI
 from fastapi_utils.tasks import repeat_every
-from news.currency_parser import get_currency_rate
 from tortoise.contrib.fastapi import register_tortoise
+from v1.auth.utils import create_root_user
+from v1.news.currency_parser import get_currency_rate
 
-from . import routers, utils
+from . import routers
 
 settings: proj_settings.Settings = proj_settings.get_settings()
 
@@ -28,7 +29,7 @@ app: FastAPI = _init_app()
 
 @app.on_event("startup")
 async def create_init_db_data():
-    await utils.create_root_user()
+    await create_root_user()
 
 
 @app.get("/")
