+++ api/v1/auth/exceptions.py
