+++ api/v1/auth/models.py
@@ -19,7 +19,7 @@ class User(base.BaseModel, mixins.TimeStamp):
     last_name = fields.CharField(max_length=50, null=True)
     password_hash = fields.CharField(max_length=128)
     roles: fields.ManyToManyRelation["Role"] = fields.ManyToManyField(
-        "models.Role", related_name="users", through="auth__user_role"
+        "models.Role", related_name="users", through="v1_auth__user_role"
     )
     disabled = fields.BooleanField(default=False)
