+++ api/v1/news/currency_parser.py
