+++ api/v1/news/schemas.py
