+++ docker-compose.yml
@@ -28,6 +28,7 @@ services:
 
   db:
     image: postgres:15.1-alpine
+    container_name: db
     restart: on-failure
     env_file:
       - ./api/.env
